requirements.txt
rumps>=0.4.0; sys_platform == 'darwin'
pystray>=0.19.0; sys_platform != 'darwin'
Pillow>=10.0.0
